Replace streaming query API with promise-based API
Previously the `Client.queries.execute` method would return a node
`Stream` that would emit Flux response data line by line. This caused
performance issues in the InfluxDB UI (see [0]).

This commit replaces the stream-based API with a promise-based API; the
`execute` method will now return a Promise that resolves with the query
response once the request has completed.

The streaming API was introduced initially to facilitate limiting
responses to a certain size. To maintain this functionality, the new
promise-based `execute` method will take a `limitChars` option (default
infinity). The response will be polled as it arrives, and if it exceeds
the character limit then the promise will reject.

Internally, the response limiting code uses `XMLHttpRequest` for
cross-browser support. This makes the client library work for browser JS
only, which is typical of similar client libraries.

[0]: https://github.com/influxdata/influxdb/issues/14084

Co-authored-by: Andrew Watkins <andrew.watkinz@gmail.com>